% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ts.R
\name{as_ts}
\alias{as_ts}
\title{Convert Spatial Raster Data to Time Series}
\usage{
as_ts(
  x,
  detrend = F,
  k = 2,
  hour2day_fun = "mean",
  domain = NULL,
  sf_obj = NULL
)
}
\arguments{
\item{x}{A file path to a NetCDF file or a `SpatRaster` object.}

\item{detrend}{Logical; should the data be detrended? Defaults to FALSE.}

\item{k}{Degree of the polynomial for detrending if `detrend` is TRUE.}

\item{hour2day_fun}{A character string specifying the function to aggregate hourly data to daily. Default is "mean".}

\item{domain}{Optional; a spatial object or extent to restrict the analysis within a specific domain.}

\item{sf_obj}{Optional; an `sf` object used to extract data points based on spatial features.}
}
\value{
A `tibble` containing time series data with two columns: time and the variable of interest.
}
\description{
This function processes a spatial raster dataset, optionally detrending and aggregating by time,
to convert it into a time series format. It supports conversion from hourly to daily data and can
apply a domain or spatial filter using an `sf` object.
}
\examples{
\dontrun{
# Assuming 'data.nc' contains a SpatRaster with time variable
raster_path <- system.file("extdata", "data.nc", package = "yourPackageName")
raster_data <- rast(raster_path)
time_series_data <- as_ts(raster_data, detrend = TRUE, k = 2, hour2day_fun = "mean")}
}
