% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_spatanalogs.R
\name{bs_spatanalogs}
\alias{bs_spatanalogs}
\title{Bootstrapping Spatial Analogs}
\usage{
bs_spatanalogs(
  x,
  analogs,
  n = 1000,
  event_fun = "mean",
  anom = NULL,
  ref_period = NULL,
  replace = TRUE,
  detrend = FALSE,
  k = 2
)
}
\arguments{
\item{x}{A file path to a NetCDF file or a `SpatRaster` object.}

\item{analogs}{A data frame containing analog dates and corresponding period labels.}

\item{n}{The number of bootstrap samples to generate.}

\item{event_fun}{A character string specifying the aggregation function to apply, default is "mean".}

\item{anom}{Logical; if TRUE, calculates anomalies based on a reference period.}

\item{ref_period}{A numeric vector of length two specifying the start and end years for the reference period, required if `anom` is TRUE.}

\item{replace}{Logical; should sampling be with replacement? Defaults to TRUE.}

\item{detrend}{Logical; should the data be detrended? Defaults to FALSE.}

\item{k}{Degree of the polynomial for detrending if `detrend` is TRUE.}
}
\value{
A list of `SpatRaster` objects for each unique period in the analogs, representing simulated bootstrap samples.
}
\description{
This function performs bootstrapping on spatial analogs for a given set of event dates.
It allows for detrending, anomaly calculation, and custom event functions on a `SpatRaster`.
The function returns a set of spatial datasets simulating possible scenarios.
}
\examples{
dat_path <- system.file("extdata", "example.nc", package = "yourPackageName")
analogs_data <- data.frame(time = as.Date('2021-01-01') + 0:10, period = rep('2021', 11))
result <- bs_spatanalogs(dat_path, analogs_data, n = 100, event_fun = "mean", anom = TRUE, ref_period = c(1980, 2000))
}
