% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_temperature.R
\name{downscale_temperature}
\alias{downscale_temperature}
\title{Downscale Temperature Using Spatial Analogs}
\usage{
downscale_temperature(x, analogs = analogs, dem = NULL, disagg, z = 6)
}
\arguments{
\item{x}{A list of raster layers or a SpatRaster object representing different periods or scenarios.}

\item{analogs}{A list specifying the time periods for the analogs used in downscaling.}

\item{dem}{An optional digital elevation model raster. If not provided, it will be automatically
downloaded using the `elevatr` package.}

\item{disagg}{A numeric value indicating the disaggregation factor for downscaling.}

\item{z}{An integer specifying the zoom level for DEM retrieval if `dem` is not provided.}
}
\value{
A raster object representing the downscaled temperature data across different periods.
}
\description{
This function downscales temperature data using spatial analogs. It integrates various
geographic and statistical methodologies to refine coarse-resolution temperature data
to a finer resolution based on digital elevation models (DEM) and other geographical
transformations. The function uses Kriging for spatial interpolation.
}
\examples{
\dontrun{
# Assuming 'temp_rasters' is a list of raster layers:
# temp_rasters <- list(rast1, rast2, rast3)
# analogs <- list(time = c("2001-01-01", "2001-01-02", "2001-01-03"))
# dem_raster <- rast("path_to_dem.tif")
# downscaled_temps <- downscale_temperature(temp_rasters, analogs = analogs, dem = dem_raster, disagg = 2, z = 6)}
}
